Extracted output functions to utils.py
